implementing get_state request, adding status to response, extracting to objects
